+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Add an opt-in `showInjectedTime` notification for user-visible injected timestamps.
+
 ## [Unreleased]
 
 ### Added